fix: APIGatewayのメソッドを一律でPOSTに変更
